libm:atan2: Adding optimized C implementation of atan2
Changed data types from unsigned long long to uint64_t in reference
implementations.

AMD-Internal: [CPUPL-2083]

+++ src/ref/atan2.c
@@ -541,7 +541,7 @@ double ALM_PROTO_REF(atan2)(double y, double x)
 
   /* Find properties of arguments x and y. */
 
-  unsigned long long ux, ui, aux, xneg, uy, auy, yneg;
+  uint64_t ux, ui, aux, xneg, uy, auy, yneg;
 
   GET_BITS_DP64(x, ux);
   GET_BITS_DP64(y, uy);
